requirements.txt
aiofiles>=0.8.0
aiohttp>=3.8.0
anthropic
asyncio-mqtt
docling
google-genai
mcp-agent
mcp-server-git
nest_asyncio
openai
pathlib2
PyPDF2>=2.0.0
reportlab>=3.5.0
streamlit
httpx>=0.24.0  
lxml>=4.9.0  
python-dotenv>=0.19.0
